requirements.txt
# PhotoPy Pro - Editor Avanzado de Imágenes en Python
# Dependencias requeridas

# Procesamiento numérico y científico
numpy>=1.21.0

# Procesamiento de imágenes
opencv-python>=4.5.0
Pillow>=8.0.0

# Interfaz gráfica
PyQt6>=6.0.0

# Librerías estándar de Python (no requieren instalación)
# sys, os, io, math, random - incluidas en Python